% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcltk2-package.R
\docType{package}
\name{tcltk2-package}
\alias{tcltk2}
\alias{tcltk2-package}
\title{tcltk2: Tcl/Tk Additions}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A series of additional Tcl commands and Tk widgets to supplement the tcltk package.
}
\details{
The \{tcltk2\} package provides additions to the \{tcltk\} base package.

Many Tk widgets are added. Also, features of the Tcl language that can be
useful also in R are made accessible through R functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SciViews/tcltk2/}
  \item \url{https://www.sciviews.org/tcltk2/}
  \item \url{https://sciviews.r-universe.dev/tcltk2}
  \item Report bugs at \url{https://github.com/SciViews/tcltk2/issues/}
}

}
\author{
\strong{Maintainer}: Philippe Grosjean \email{phgrosjean@sciviews.org} (\href{https://orcid.org/0000-0002-2694-9471}{ORCID})

}
\keyword{internal}
