\name{tcltk2-package}
\alias{tcltk2-package}
\alias{tcltk2}

\docType{package}

\title{
Tcl/Tk Additions
}

\description{
Additions to the tcltk package: more Tk widgets and Tcl/Tk commands.
}

\details{
A series of additional Tk widgets with style and various functions
to supplement the tcltk package.
#ifdef windows
(under Windows: DDE exchange, access to the registry and icon manipulation)
#endif
 
For a complete list of functions, use \code{ls("package:tcltk2")}.
}

\author{
Philippe Grosjean

Maintainer: Philippe Grosjean \email{phgrosjean@sciviews.org}
}

\keyword{ package }
