\name{tcltk2-package}
\alias{tcltk2-package}
\alias{tcltk2}

\docType{package}

\title{
Tcl/Tk Additions
}

\description{
Additions to the tcltk package: more Tk widgets and Tcl/Tk commands.
}

\details{
\tabular{ll}{
Package: \tab tcltk2\cr
Type: \tab Package\cr
Version: \tab 1.2-5\cr
Date: \tab 2013-03-10\cr
License: \tab LGPL-3 plus see LICENSE file\cr
LazyLoad: \tab yes\cr
}
A series of additional Tk widgets with style and various functions
to supplement the tcltk package.
#ifdef windows
(under Windows: DDE exchange, access to the registry and icon manipulation)
#endif
}

\author{
Philippe Grosjean

Maintainer: Philippe Grosjean \email{phgrosjean@sciviews.org}
}

\seealso{ \code{\link{tk2widgets}}, \code{\link{tclFun}} }

\keyword{ package }

\concept{ more Tcl/Tk commands and widgets }
