% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tk2methods.R
\name{is.tk2widget}
\alias{is.tk2widget}
\alias{print.tk2widget}
\alias{tk2cfglist}
\alias{print.tk2cfglist}
\alias{state}
\alias{state.tk2widget}
\alias{label}
\alias{label.tk2widget}
\alias{label<-}
\alias{label<-.tk2widget}
\alias{tag}
\alias{tag.tk2widget}
\alias{tag<-}
\alias{tag<-.tk2widget}
\alias{disabled}
\alias{disabled.tk2widget}
\alias{disabled<-}
\alias{disabled<-.tk2widget}
\alias{values}
\alias{values.tk2widget}
\alias{values.tk2listbox}
\alias{values<-}
\alias{values<-.tk2widget}
\alias{values<-.tk2listbox}
\alias{value}
\alias{value.tk2widget}
\alias{value.tk2listbox}
\alias{value<-}
\alias{value<-.tk2widget}
\alias{value<-.tk2listbox}
\alias{selection}
\alias{selection.tk2widget}
\alias{selection.tk2listbox}
\alias{selection<-}
\alias{selection<-.tk2widget}
\alias{selection<-.tk2listbox}
\alias{visibleItem}
\alias{visibleItem.tk2widget}
\alias{visibleItem.tk2listbox}
\alias{size}
\alias{size.tk2widget}
\alias{size.tk2listbox}
\alias{config}
\alias{config.tk2widget}
\alias{config.tk2label}
\alias{config<-}
\alias{config<-.tk2widget}
\alias{config<-.tk2label}
\title{A series of methods applicable to tk2widget or tk2cfglist objects}
\usage{
is.tk2widget(x)

\method{print}{tk2widget}(x, ...)

tk2cfglist(...)

\method{print}{tk2cfglist}(x, ...)

state(x, ...)

\method{state}{tk2widget}(x, ...)

label(x, ...)

\method{label}{tk2widget}(x, ...)

label(x) <- value

\method{label}{tk2widget}(x) <- value

tag(x, ...)

\method{tag}{tk2widget}(x, ...)

tag(x) <- value

\method{tag}{tk2widget}(x) <- value

disabled(x, ...)

\method{disabled}{tk2widget}(x, ...)

disabled(x) <- value

\method{disabled}{tk2widget}(x) <- value

values(x, ...)

\method{values}{tk2widget}(x, ...)

\method{values}{tk2listbox}(x, ...)

values(x) <- value

\method{values}{tk2widget}(x) <- value

\method{values}{tk2listbox}(x) <- value

value(x, ...)

\method{value}{tk2widget}(x, ...)

\method{value}{tk2listbox}(x, ...)

value(x) <- value

\method{value}{tk2widget}(x) <- value

\method{value}{tk2listbox}(x) <- value

selection(x, ...)

\method{selection}{tk2widget}(x, ...)

\method{selection}{tk2listbox}(x, ...)

selection(x) <- value

\method{selection}{tk2widget}(x) <- value

\method{selection}{tk2listbox}(x) <- value

visibleItem(x, index, ...)

\method{visibleItem}{tk2widget}(x, index, ...)

\method{visibleItem}{tk2listbox}(x, index, ...)

size(x, ...)

\method{size}{tk2widget}(x, ...)

\method{size}{tk2listbox}(x, ...)

config(x, ...)

\method{config}{tk2widget}(x, cfglist, ...)

\method{config}{tk2label}(x, cfglist, ...)

config(x) <- value

\method{config}{tk2widget}(x) <- value

\method{config}{tk2label}(x) <- value
}
\arguments{
\item{x}{A tk2widget object.}

\item{...}{A series of named arguments corresponding to parameters and values
to use for the configuration for \code{tk2cfglist()}, or reserved arguments for
future use for the other function (not used yet).}

\item{value}{A value to assign to the object's method.}

\item{index}{The zero-based index of the item to make visible.}

\item{cfglist}{a list containing one or more named items, with the name
being a Tcl/Tk property and items being the new value for the property.}
}
\value{
Depends on the function. The \code{is.xxx()} function return \code{TRUE} or \code{FALSE} if
the object is of the right class or not. The assignations form return the
assigned value. The direct form return the item.
}
\description{
Tk2widgets can be used as tcltk widgets, but they propose also an
object-oriented interaction through these different methods.
}
\seealso{
\code{\link[=tk2button]{tk2button()}}, \code{\link[=tk2tip]{tk2tip()}}
}
\author{
Philippe Grosjean
}
\keyword{utilities}
